fastapi/requirements.txt
fastapi==0.85.1
orjson==3.8.0
pydantic[dotenv]==1.10.2
uvicorn==0.19.0
uvloop==0.17.0
aiohttp==3.8.3
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-instrumentation-fastapi==0.35b0
opentelemetry-exporter-jaeger==1.14.0